Refactor: refactor overall components
styled-component 부 분 코드 하단으로 이동

.jsx 파일 전부 .js파일로 통일

Modal/Button 세 버튼 ModalButton으로 통합, common/Button도 ModalButton에 통합

 ModalInfo 컴포넌트로  분리, HeaderMenu 컴포넌트로 분리
