+++ src/js/components/Modal/ModalContent.js
@@ -1,5 +1,14 @@
-
 import styled from "styled-components";
+
+const Content = ({ title, body }) => {
+	const free = "(40,000원 이상 구매 시 무료)";
+	return (
+		<ContentWrapper>
+			<ContentTitle>{title}</ContentTitle>
+			<ContentBody dangerouslySetInnerHTML={{__html:body.replace(free, `<b>${free}</b>`)}} />
+		</ContentWrapper>
+	);
+};
 
 const ContentWrapper = styled.div`
 	display: flex;
@@ -20,17 +29,5 @@ const ContentBody = styled.div`
 	width: 364px;
 	color: #4f4f4f;
 `;
-const Content = ({ title, body }) => {
-	const free = "(40,000원 이상 구매 시 무료)";
-	return (
-		<ContentWrapper>
-			<ContentTitle>{title}</ContentTitle>
-			<ContentBody>
-				{body.replace(free, "")}
-				<b>{body.includes(free) ? free : ""}</b>
-			</ContentBody>
-		</ContentWrapper>
-	);
-};
 
 export default Content
