+++ src/js/components/Modal/ModalPrice.js
@@ -1,6 +1,14 @@
 import styled from "styled-components";
 import Badge from "../common/Badge"
 
+const Price = ({data}) => (
+	<PriceWrapper>
+		{data.badge && <Badge badge={data.badge} />}
+		<SaledPrice>{data.prices[1] || data.prices[0]}</SaledPrice>
+		<OriginPrice>{data.prices[1] && data.prices[0]}</OriginPrice>
+	</PriceWrapper>
+);
+
 const PriceWrapper = styled.div`
 	display: flex;
 	flex-direction: row;
@@ -21,13 +29,5 @@ const OriginPrice = styled.div`
 	text-decoration-line: line-through;
 	color: #828282;
 `;
-
-const Price = ({data}) => (
-	<PriceWrapper>
-		{!data.badge || <Badge badge={data.badge} />}
-		<SaledPrice>{data.prices[1] ? data.prices[1] : data.prices[0]}</SaledPrice>
-		<OriginPrice>{!data.prices[1] || data.prices[0]}</OriginPrice>
-	</PriceWrapper>
-);
 
 export default Price
