+++ src/js/components/Modal/ModalThumbNail.js
@@ -1,34 +1,6 @@
 import { useEffect, useState } from "react";
 import styled from "styled-components";
 
-const ThumbNailWrapper = styled.div`
-	position: relative;
-	width: 392px;
-	height: 472px;
-	left: 48px;
-	top: 48px;
-`;
-const TopImage = styled.img`
-	width: 392px;
-	height: 392px;
-	margin-bottom: 8px;
-`;
-
-const ThumbImages = styled.ul`
-	display: flex;
-	flex-direction: row;
-	justify-content: space-between;
-`;
-const ThumbImage = styled.li`
-	width: 72px;
-	height: 72px;
-	img {
-		width: 100%;
-		height: 100%;
-	}
-	background: #f5f5f7;
-	border-radius: 5px;
-`;
 const ThumbNail = (props) => {
 	const [topImage, setTopImage] = useState(props.top_image);
 	const clickThumbNail = ({ target }) => setTopImage(() => target.currentSrc);
@@ -44,5 +16,33 @@ const ThumbNail = (props) => {
 		</ThumbNailWrapper>
 	);
 };
+
+const ThumbNailWrapper = styled.div`
+	position: relative;
+	width: 392px;
+	height: 472px;
+	left: 48px;
+	top: 48px;
+`;
+const TopImage = styled.img`
+	width: 392px;
+	height: 392px;
+	margin-bottom: 8px;
+`;
+const ThumbImages = styled.ul`
+	display: flex;
+	flex-direction: row;
+	justify-content: space-between;
+`;
+const ThumbImage = styled.li`
+	width: 72px;
+	height: 72px;
+	img {
+		width: 100%;
+		height: 100%;
+	}
+	background: #f5f5f7;
+	border-radius: 5px;
+`;
 
 export default ThumbNail;
